requirements.txt
streamlit>=1.34.0
python-dotenv>=1.0.1
google-adk>=0.2.0
google-genai>=0.2.2